add Category service, repositary and sql schema

+++ ewm-service/src/main/java/ru/practicum/ewm/dto/EventRequestStatusUpdateRequest.java
@@ -19,7 +19,7 @@ public class EventRequestStatusUpdateRequest {
     /**
      * Идентификаторы запросов на участие в событии текущего пользователя
      */
-    private Set<Integer> requestIds;
+    private Set<Long> requestIds;
     /**
      * Новый статус запроса на участие в событии текущего пользователя
      */
